Step 1: Refactor, moving departures.json fetchers to TripKit
