+++ Sources/TripKit/model/API/DeparturesAPIModel.swift
@@ -9,6 +9,7 @@ import Foundation
 
 extension TKAPI {
 
+  /// Output data model for `departures.json`
   public struct Departures: Codable, Hashable {
     
     public struct Embarkations: Codable, Hashable {
